backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
redis==5.0.1
qdrant-client==1.7.0
scikit-surprise==1.1.3
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
httpx==0.26.0
groq==0.4.1
pillow==10.2.0
imagehash==4.3.1
boto3==1.34.34
lime==0.2.0.1
prometheus-client==0.19.0
structlog==24.1.0
python-jose[cryptography]==3.3.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
